# ⚠️ Change client.Patch to take client.Object for performance