--- a/pkg/client/patch.go
+++ b/pkg/client/patch.go
@@ -20,8 +20,6 @@ import (
 	"fmt"
 
 	jsonpatch "github.com/evanphx/json-patch"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/apimachinery/pkg/util/json"
@@ -47,7 +45,7 @@ func (s *patch) Type() types.PatchType {
 }
 
 // Data implements Patch.
-func (s *patch) Data(obj runtime.Object) ([]byte, error) {
+func (s *patch) Data(obj Object) ([]byte, error) {
 	return s.data, nil
 }
 
@@ -87,7 +85,7 @@ type MergeFromOptions struct {
 }
 
 type mergeFromPatch struct {
-	from runtime.Object
+	from Object
 	opts MergeFromOptions
 }
 
@@ -96,8 +94,12 @@ func (s *mergeFromPatch) Type() types.PatchType {
 	return types.MergePatchType
 }
 
-func (*mergeFromPatch) data(original, modified Object, opts MergeFromOptions) ([]byte, error) {
-	if opts.OptimisticLock {
+// Data implements Patch.
+func (s *mergeFromPatch) Data(obj Object) ([]byte, error) {
+	original := s.from
+	modified := obj
+
+	if s.opts.OptimisticLock {
 		version := original.GetResourceVersion()
 		if len(version) == 0 {
 			return nil, fmt.Errorf("cannot use OptimisticLock, object %q does not have any resource version we can use", original)
@@ -120,64 +122,21 @@ func (*mergeFromPatch) data(original, modified Object, opts MergeFromOptions) ([
 		return nil, err
 	}
 
-	return jsonpatch.CreateMergePatch(originalJSON, modifiedJSON)
-}
-
-// Data implements Patch.
-func (s *mergeFromPatch) Data(obj runtime.Object) ([]byte, error) {
-	fromObject, fromOk := s.from.(Object)
-	objObject, objOk := obj.(Object)
-
-	if fromOk && objOk {
-		return s.data(fromObject, objObject, s.opts)
-	}
-
-	originalJSON, err := json.Marshal(s.from)
-	if err != nil {
-		return nil, err
-	}
-
-	modifiedJSON, err := json.Marshal(obj)
-	if err != nil {
-		return nil, err
-	}
-
 	data, err := jsonpatch.CreateMergePatch(originalJSON, modifiedJSON)
 	if err != nil {
 		return nil, err
-	}
-
-	if s.opts.OptimisticLock {
-		dataMap := map[string]interface{}{}
-		if err := json.Unmarshal(data, &dataMap); err != nil {
-			return nil, err
-		}
-		fromMeta, ok := s.from.(metav1.Object)
-		if !ok {
-			return nil, fmt.Errorf("cannot use OptimisticLock, from object %q is not a valid metav1.Object", s.from)
-		}
-		resourceVersion := fromMeta.GetResourceVersion()
-		if len(resourceVersion) == 0 {
-			return nil, fmt.Errorf("cannot use OptimisticLock, from object %q does not have any resource version we can use", s.from)
-		}
-		u := &unstructured.Unstructured{Object: dataMap}
-		u.SetResourceVersion(resourceVersion)
-		data, err = json.Marshal(u)
-		if err != nil {
-			return nil, err
-		}
 	}
 
 	return data, nil
 }
 
 // MergeFrom creates a Patch that patches using the merge-patch strategy with the given object as base.
-func MergeFrom(obj runtime.Object) Patch {
+func MergeFrom(obj Object) Patch {
 	return &mergeFromPatch{from: obj}
 }
 
 // MergeFromWithOptions creates a Patch that patches using the merge-patch strategy with the given object as base.
-func MergeFromWithOptions(obj runtime.Object, opts ...MergeFromOption) Patch {
+func MergeFromWithOptions(obj Object, opts ...MergeFromOption) Patch {
 	options := &MergeFromOptions{}
 	for _, opt := range opts {
 		opt.ApplyToMergeFrom(options)