@vincepri 2021-03-01
/milestone v0.9.x

If we switch the interface accepted in the method, we need to mark this as ⚠️