@cbandy 2021-03-02
If we're changing the signatures, perhaps we should add options vararg to MergeFrom like other patch constructors.