@cbandy 2021-03-02
> @cbandy Could we discuss that change in a separate issue?

Certainly. Is the signature (with runtime.Object) acceptable as-is in this PR?