@vincepri 2021-03-02
We should move to `client.Object` and always use the fast path, there is no need to use the old runtime.Object